+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased - ReleaseDate
+
+## 0.3.0 - 2021-01-31
 ### Added
 - The `context` module.